feat(cmd): use log.Fatalf for ExitOnError
On failures we would not see what the error is.
